MapController methods. State methods. Code cleanup
In as many places as I could, I made the MapController use the state. I also filled out the relevant methods for the states. It appears as if we are missing some methods for our states. I'll worry about that later.
I also tried to clean up the code as much as possible.